requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
anthropic>=0.18.0
gunicorn>=21.2.0

